+++ api/models/single_cell.py
@@ -1,4 +1,4 @@
-from api.base import db
+from api import db
 
 
 class SingleCell(db.Model):
