+++ app.py
@@ -1,4 +1,4 @@
-from api.base import app
+from api import app
 
 if __name__ == '__main__':
     app.run()
